Added tests, improved routing for BT
