+++ templates/braintree/index.html.twig
@@ -19,7 +19,7 @@
     </div>
     <div class="row">
         <div class="col-12 mt-5 text-center">
-            <a class="btn btn-dark bg-braintree" href="{{ path('braintree-payments') }}">Try out Payments!</a>
+            <a class="btn btn-dark bg-braintree" href="{{ path('braintree-payments-index') }}">Try out Payments!</a>
             <h6 class="ml-5 d-inline bullet">Test payments with a DropUI format. Using latest version of JS SDK v3</h6>
         </div>
     </div>
