+++ templates/navbar/dropdown.html.twig
@@ -6,12 +6,10 @@
         <div class="dropdown-menu" aria-labelledby="navbarDropdown">
             <a class="dropdown-item bullet" href="{{ path('index') }}">PlayGround Home</a>
             <div class="dropdown-divider"></div>
-            <h6 class="dropdown-header">PayPal Commerce Platform </h6><p class="red-text  d-inline-block">Business</p>
-            <a class="dropdown-item bullet" href="{{ path('paypal-index') }}">Connect with PayPal</a>
-            <a class="dropdown-item bullet disabled" href="#">Make Payments</a>
-            <a class="dropdown-item bullet disabled" href="#">Receive Payments</a>
+            <h6 class="dropdown-header d-inline">PayPal Commerce Platform <strong class="red-text">Business</strong></h6>
+            <a class="dropdown-item bullet" href="{{ path('paypal-index') }}">PayPal Checkout</a>
             <div class="dropdown-divider"></div>
-            <h6 class="dropdown-header">PayPal Commerce Platform </h6><p class="red-text  d-inline-block">Enterprise</p>
+            <h6 class="dropdown-header d-inline">PayPal Commerce Platform <strong class="red-text">Enterprise</strong></h6>
             <a class="dropdown-item bullet" href="{{ path('braintree-index') }}">BrainTree</a>
             <a class="dropdown-item bullet" href="#">HyperWallet</a>
         </div>
