+++ tests/Controller/Braintree/DefaultControllerTest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Tests\Controller\Braintree;
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+/**
+ * Class DefaultControllerTest
+ * @package App\Tests\Controller\Braintree
+ */
+class DefaultControllerTest extends WebTestCase
+{
+    /**
+     * testIndex
+     */
+    public function testIndex()
+    {
+        $client = static::createClient();
+        $container = self::$container;
+        $indexRoute = $container->get('router')->generate('braintree-index');
+        $client->request('GET', $indexRoute);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+        $this->assertStringContainsString(
+            "Let's Play",
+            $client->getResponse()->getContent()
+        );
+    }
+}
